fix: Use BlobAssetReference for nested state machines
The SubStateMachineBlob.NestedStateMachine field was incorrectly
defined as inline StateMachineBlob, but the converter was trying
to assign a BlobAssetReference to it using a non-existent method.

Fix:
- Change field to BlobAssetReference<StateMachineBlob>
- Assign reference directly in converter
- Access via .Value in runtime code

This enables proper recursive blob structures for hierarchical
state machines.

+++ Runtime/Systems/InitializeStateMachineStackSystem.cs
@@ -33,7 +33,7 @@ namespace DMotion
                 // Initialize with root context
                 buffer.Add(new StateMachineContext
                 {
-                    CurrentStateIndex = stateMachine.ValueRO.CurrentState.StateIndex,
+                    CurrentStateIndex = (short)stateMachine.ValueRO.CurrentState.StateIndex,
                     ParentSubMachineIndex = -1, // Root level
                     Level = 0
                 });
